Renamed enums.py to messages.py
